refactor: split service for better separation of concerns
